Use TimeNano for all timings

+++ core/src/mindmelt/game/spells/TimedSpell.java
@@ -3,14 +3,14 @@ package mindmelt.game.spells;
 import mindmelt.game.engine.Engine;
 
 public class TimedSpell extends Spell {
-    protected float expiry;
+    protected long expiry;
 
     @Override
     public void update(Engine engine) {
         setActive(engine.getSystemTime()<expiry);
     }
 
-    protected void init(Engine engine, float time) {
+    protected void init(Engine engine, long time) {
         expiry = engine.getSystemTime()+time;
     }
 }
